requirements.txt
asgiref==3.8.1
certifi==2024.7.4
Django==5.0.7
django-rest-knox==5.0.1
djangorestframework==3.15.2
djangorestframework-gis==1.0
geographiclib==2.0
geopy==2.4.1
haversine==2.8.1
mysqlclient==2.2.4
numpy==2.0.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyproj==3.6.1
shapely==2.0.5
sqlparse==0.5.1
tzdata==2024.1
